requirements.txt
requests
numpy
pandas
openai
fastapi
arxiv
python-dotenv
pydantic
uvicorn
